Removing event handling from ethpy, and moving db specific event getters to chainsync
